\name{data.Y}
\alias{data.Y}
%- Also NEED an '\alias' for EACH other topic documented here.
\docType{data}
\title{
Data set for ZIBBSeqDiscovery
}
\description{
Data set for ZIBBSeqDiscovery. The data set data.Y is a count matrix with 100 OTUs and 20 samples randomly selected from a real microbiome data set.
}
\usage{
data(data.Y)
}
%- maybe also 'usage' for other objects documented here.
%\author{
%Tao Hu, Yihui Zhou
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
## Load the data
data(data.Y)
dim(data.Y)
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
