\name{kostic.x}
\alias{kostic.x}
\docType{data}
\title{
Data set for ZIBBSeqDiscovery
}
\description{
Data set for ZIBBSeqDiscovery. The data set kostic.x is a count matrix with 2505 OTUs and 185 samples randomly selected from a real microbiome data set kostic.
}
\usage{data("kostic.x")}
\references{
Kostic, A. D., Gevers, D., Pedamallu, C. S., Michaud, M., Duke, F., Earl, A. M., Ojesina, A. I., Jung, J., Bass, A. J., Tabernero, J., et al. (2012). Genomic analysis identifies association of fusobacterium with colorectal carcinoma. Genome research 22, 292-298.
}
\examples{
## load the data
data(kostic.x)
dim(kostic.x)
}
%\keyword{datasets}
