\name{kostic.y}
\alias{kostic.y}
\docType{data}
\title{
Data set for ZIBBSeqDiscovery
}
\description{
Data set for ZIBBSeqDiscovery. The data set kostic.y is the phenotype of the 185 samples from a real microbiome data set kostic.
}
\usage{data("kostic.y")}
\references{
Kostic, A. D., Gevers, D., Pedamallu, C. S., Michaud, M., Duke, F., Earl, A. M., Ojesina, A. I., Jung, J., Bass, A. J., Tabernero, J., et al. (2012). Genomic analysis identifies association of fusobacterium with colorectal carcinoma. Genome research 22, 292-298.
}
\examples{
## load the data
data(kostic.y)
length(kostic.y)
}
%\keyword{datasets}
