\name{ZIBBSeqDiscovery-package}
\alias{ZIBBSeqDiscovery-package}
\alias{ZIBBSeqDiscovery}
\docType{package}
\title{
\packageTitle{ZIBBSeqDiscovery}
}
\description{
\packageDescription{ZIBBSeqDiscovery}
}
\details{
%The DESCRIPTION file:
%\packageDESCRIPTION{ZIBBSeqDiscovery}
\packageIndices{ZIBBSeqDiscovery}
}
\author{
\packageAuthor{ZIBBSeqDiscovery}

Maintainer: \packageMaintainer{ZIBBSeqDiscovery}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
